rag_core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-core"
version = "0.1.0"
description = "可移植的 RAG (Retrieval-Augmented Generation) 模組"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "sglangRAG Team"}
]
keywords = ["rag", "retrieval", "embeddings", "llm", "chromadb", "bm25"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "chromadb>=0.4.0",
    "rank-bm25>=0.2.2",
    "jieba>=0.42.1",
    "FlagEmbedding>=1.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
pdf = ["pymupdf>=1.23.0"]
docx = ["python-docx>=1.0.0"]
full = [
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/monkey1sai/sglangRAG"
Repository = "https://github.com/monkey1sai/sglangRAG"

[project.scripts]
rag-server = "rag_core.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["rag_core*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
